pyproject.toml
[tool.poetry]
name = "2020-pcs3216-9345853"
version = "0.1.0"
description = "Projeto de SisProg"
authors = ["Adarah <lucasyharada@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.4.1"
attr = "^0.3.1"
numpy = "^1.18.3"
pyparsing = "^2.4.7"
attrs = "^19.3.0"
hypothesis-pytest = "^0.19.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
isort = "^4.3.21"
black = "^19.10b0"
pytest = "^5.4.1"
ptvsd = "^4.3.2"
pytest-mock = "^3.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
